fix: address all PR #22 review issues
Blockers:
- #1: Extract ChatPanelHelpers.tsx + useChatPromptInput hook (ChatPanel 617→371 LOC)
- #2: Extract preload/user-feedback.ts + preload/debug-lsp.ts (preload 561→482 LOC)
- #3: Remove hardcoded sessionId from PendingQuestion type and usage
- #4: Remove phantom @mariozechner/pi-ai peer dependency

Should-fix:
- #5: Extract shared emitter.ts; IPC handler uses local mirror with doc link
- #6: Add 'source of truth' doc comments on mirrored host types (rootDir constraint)
- #7: Add TTL-based eviction (10min) for pendingQuestions Map to prevent leaks
- #8: Add proper type guard for QuestionnaireNotice tool input validation

Nits:
- #9: Replace inline import() types with top-level Theme import in tui-questionnaire
- #10: Don't auto-advance past last question in QuestionnaireForm
- #11: Add aria-label to icon-only buttons, role/tabIndex/onKeyDown to notice div
- #12: Fix comment claiming three capabilities (only two tools exist)
