+++ apps/desktop/electron/preload.ts
@@ -1,5 +1,7 @@
 import { contextBridge, ipcRenderer, type IpcRendererEvent } from 'electron';
 import { IpcChannels } from '../src/types/ipc';
+import { userFeedbackBridge } from './preload/user-feedback';
+import { debugBridge, lspBridge } from './preload/debug-lsp';
 import type {
   WorkspaceInfo,
   WorkspaceConfig,
@@ -432,45 +434,7 @@ contextBridge.exposeInMainWorld('sero', {
       ipcRenderer.invoke(IpcChannels.feedback.remove, messageId),
   },
 
-  userFeedback: {
-    /** Get all currently pending questions (for mount-time hydration). */
-    getPending: (): Promise<import('../src/types/ipc').UserFeedbackPendingQuestion[]> =>
-      ipcRenderer.invoke(IpcChannels.userFeedback.getPending),
-
-    /** Send user's answer to a pending question/questionnaire. */
-    answer: (response: import('../src/types/ipc').UserFeedbackResponse): Promise<void> => {
-      // Fire a DOM event synchronously so all renderer stores (Zustand, federated app)
-      // can clear immediately — no IPC round-trip needed.
-      window.dispatchEvent(
-        new CustomEvent('sero:user-feedback:answered', { detail: { id: response.id } }),
-      );
-      return ipcRenderer.invoke(IpcChannels.userFeedback.answer, response);
-    },
-
-    /** Listen for incoming question/questionnaire requests from extensions. */
-    onQuestion: (
-      callback: (data: import('../src/types/ipc').UserFeedbackPendingQuestion) => void,
-    ): (() => void) => {
-      const handler = (
-        _event: Electron.IpcRendererEvent,
-        data: import('../src/types/ipc').UserFeedbackPendingQuestion,
-      ) => callback(data);
-      ipcRenderer.on(IpcChannels.userFeedback.question, handler);
-      return () => {
-        ipcRenderer.removeListener(IpcChannels.userFeedback.question, handler);
-      };
-    },
-
-    /** Listen for cancellation of a pending question. */
-    onCancel: (callback: (data: { id: string }) => void): (() => void) => {
-      const handler = (_event: Electron.IpcRendererEvent, data: { id: string }) =>
-        callback(data);
-      ipcRenderer.on(IpcChannels.userFeedback.cancel, handler);
-      return () => {
-        ipcRenderer.removeListener(IpcChannels.userFeedback.cancel, handler);
-      };
-    },
-  },
+  userFeedback: userFeedbackBridge,
 
   editor: {
     readFile: (workspaceId: string, filePath: string): Promise<string> =>
@@ -511,51 +475,8 @@ contextBridge.exposeInMainWorld('sero', {
     },
   },
 
-  debug: {
-    toggle: (): Promise<boolean> =>
-      ipcRenderer.invoke(IpcChannels.debug.toggle),
+  debug: debugBridge,
 
-    getState: (): Promise<boolean> =>
-      ipcRenderer.invoke(IpcChannels.debug.getState),
-
-    openLog: (): Promise<void> =>
-      ipcRenderer.invoke(IpcChannels.debug.openLog),
-
-    clearLog: (): Promise<void> =>
-      ipcRenderer.invoke(IpcChannels.debug.clearLog),
-
-    onStateChanged: (callback: (enabled: boolean) => void): (() => void) => {
-      const handler = (_event: Electron.IpcRendererEvent, enabled: boolean) => {
-        callback(enabled);
-      };
-      ipcRenderer.on(IpcChannels.debug.stateChanged, handler);
-      return () => {
-        ipcRenderer.removeListener(IpcChannels.debug.stateChanged, handler);
-      };
-    },
-  },
-
-  lsp: {
-    start: (workspaceId: string, languageId: string) =>
-      ipcRenderer.invoke(IpcChannels.lsp.start, workspaceId, languageId),
-    stop: (workspaceId: string, language: string): Promise<void> =>
-      ipcRenderer.invoke(IpcChannels.lsp.stop, workspaceId, language),
-    request: (workspaceId: string, language: string, method: string, params?: unknown) =>
-      ipcRenderer.invoke(IpcChannels.lsp.request, workspaceId, language, method, params),
-    notify: (workspaceId: string, language: string, method: string, params?: unknown): void =>
-      ipcRenderer.send(IpcChannels.lsp.notify, workspaceId, language, method, params),
-    hasServer: (workspaceId: string, language: string): Promise<boolean> =>
-      ipcRenderer.invoke(IpcChannels.lsp.hasServer, workspaceId, language),
-    onNotification: (callback: (data: { workspaceId: string; language: string; notification: any }) => void): (() => void) => {
-      const handler = (_e: IpcRendererEvent, data: any) => callback(data);
-      ipcRenderer.on(IpcChannels.lsp.notification, handler);
-      return () => { ipcRenderer.removeListener(IpcChannels.lsp.notification, handler); };
-    },
-    onServerStopped: (callback: (data: { workspaceId: string; language: string }) => void): (() => void) => {
-      const handler = (_e: IpcRendererEvent, data: any) => callback(data);
-      ipcRenderer.on(IpcChannels.lsp.serverStopped, handler);
-      return () => { ipcRenderer.removeListener(IpcChannels.lsp.serverStopped, handler); };
-    },
-  },
+  lsp: lspBridge,
 
 });
