+++ progress.md
@@ -1,0 +1,19 @@
+# Progress: pi-user-feedback
+
+## Session 1
+
+### Analysis Complete
+- Read 3 original Pi CLI extensions (question, send-user-message, questionnaire)
+- Read apps-tutorial.md, sero-extension.ts, agent IPC, ChatPanel, ToolCallGroup
+- Read Pi SDK ExtensionAPI types (hasUI, ctx.ui.custom, sendUserMessage)
+- Confirmed ctx.hasUI is false in Sero (SDK mode, no TUI)
+- Confirmed globalThis EventEmitter pattern is viable (same Node process)
+- Identified all IPC layers: channels, handlers, preload, renderer types
+
+### All Phases Complete
+- Phase 1: Package scaffold ✓
+- Phase 2: IPC bridge (host side) ✓
+- Phase 3: Renderer store + ChatPanel card ✓
+- Phase 4: Extension (Pi extension, dual-mode) ✓
+- Phase 5: Dedicated app UI (questionnaire) ✓
+- Phase 6: Build + typecheck ✓ (pnpm build, tsc --noEmit, build-electron all pass)
